+++ pallets/anonymity-mining-claims/src/test_utils.rs
@@ -1,10 +1,292 @@
+use crate::mock::*;
 use ark_ff::{BigInteger, PrimeField};
 use webb_primitives::ElementTrait;
+use ark_serialize::CanonicalSerialize;
+use num_bigint::{BigInt, Sign};
+use frame_benchmarking::account;
+use ark_bn254::{Bn254, Fr};
+use frame_support::{assert_err, assert_ok};
+use ark_circom::{read_zkey, WitnessCalculator};
+use circom_proving::circom_from_folder;
+use ark_groth16::ProvingKey;
+use ark_relations::r1cs::ConstraintMatrices;
+use webb_primitives::webb_proposals::{
+	ResourceId, SubstrateTargetSystem, TargetSystem, TypedChainId,
+};
+
+use std::{
+	fs::File,
+	str::FromStr,
+	convert::TryInto,
+	sync::Mutex,
+};
+
+use arkworks_setups::{common::setup_params, Curve};
 
 use crate::mock::Element;
 
 type Bn254Fr = ark_bn254::Fr;
-type Bn254 = ark_bn254::Bn254;
+
+const SEED: u32 = 0;
+
+
+pub fn setup_environment() {
+	let curve = Curve::Bn254;
+	let params3 = setup_params::<ark_bn254::Fr>(curve, 5, 3);
+
+	for account_id in [
+		account::<AccountId>("", 1, SEED),
+		account::<AccountId>("", 2, SEED),
+		account::<AccountId>("", 3, SEED),
+		account::<AccountId>("", 4, SEED),
+		account::<AccountId>("", 5, SEED),
+	] {
+		assert_ok!(Balances::set_balance(RuntimeOrigin::root(), account_id, 100_000_000, 0));
+		assert_ok!(HasherPallet::force_set_parameters(
+			RuntimeOrigin::root(),
+			params3.to_bytes().try_into().unwrap(),
+		));
+	}
+}
+pub fn setup_environment_with_circom(
+) -> ((ProvingKey<Bn254>, ConstraintMatrices<Fr>), &'static Mutex<WitnessCalculator>) {
+	setup_environment();
+
+	println!("Setting up ZKey");
+	let path_2_2 = "../../solidity-fixtures/solidity-fixtures/reward_2/30/circuit_final.zkey";
+	let mut file_2_2 = File::open(path_2_2).unwrap();
+	let params_2_2 = read_zkey(&mut file_2_2).unwrap();
+
+	let wasm_2_2_path = "../../solidity-fixtures/solidity-fixtures/reward_2/30/reward_30_2.wasm";
+
+	let wc_2_2 = circom_from_folder(wasm_2_2_path);
+
+	println!("Setting up the verifier pallet");
+	let mut vk_2_2_bytes = Vec::new();
+	params_2_2.0.vk.serialize(&mut vk_2_2_bytes).unwrap();
+
+	let param_call = ClaimsVerifier::force_set_parameters(
+		RuntimeOrigin::root(),
+		2,
+		vk_2_2_bytes.try_into().unwrap(),
+	);
+	assert_ok!(param_call);
+
+
+	(params_2_2, wc_2_2)
+}
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+// #[serde(rename_all = "camelCase")]
+pub struct InputsRaw {
+	pub rate: String,
+	pub fee: String,
+	#[serde(rename = "rewardNullifier")]
+	pub reward_nullifier: String,
+	#[serde(rename = "extDataHash")]
+	pub ext_data_hash: String,
+	#[serde(rename = "noteChainID")]
+	pub note_chain_id: String,
+	#[serde(rename = "noteAmount")]
+	pub note_amount: String,
+	#[serde(rename = "noteAssetID")]
+	pub note_asset_id: String,
+	#[serde(rename = "noteTokenID")]
+	pub note_token_id: String,
+	#[serde(rename = "note_ak_X")]
+	pub note_ak_x: String,
+	#[serde(rename = "note_ak_Y")]
+	pub note_ak_y: String,
+	#[serde(rename = "noteBlinding")]
+	pub note_blinding: String,
+	#[serde(rename = "notePathElements")]
+	pub note_path_elements: Vec<String>,
+	#[serde(rename = "notePathIndices")]
+	pub note_path_indices: String,
+	pub note_alpha: String,
+	#[serde(rename = "note_ak_alpha_X")]
+	pub note_ak_alpha_x: String,
+	#[serde(rename = "note_ak_alpha_Y")]
+	pub note_ak_alpha_y: String,
+	#[serde(rename = "inputChainID")]
+	pub input_chain_id: String,
+	#[serde(rename = "inputAmount")]
+	pub input_amount: String,
+	#[serde(rename = "inputPrivateKey")]
+	pub input_private_key: String,
+	#[serde(rename = "inputBlinding")]
+	pub input_blinding: String,
+	#[serde(rename = "inputNullifier")]
+	pub input_nullifier: String,
+	#[serde(rename = "inputRoot")]
+	pub input_root: String,
+	#[serde(rename = "inputPathElements")]
+	pub input_path_elements: Vec<String>,
+	#[serde(rename = "inputPathIndices")]
+	pub input_path_indices: String,
+	#[serde(rename = "outputChainID")]
+	pub output_chain_id: String,
+	#[serde(rename = "outputAmount")]
+	pub output_amount: String,
+	#[serde(rename = "outputPrivateKey")]
+	pub output_private_key: String,
+	#[serde(rename = "outputBlinding")]
+	pub output_blinding: String,
+	#[serde(rename = "outputCommitment")]
+	pub output_commitment: String,
+	#[serde(rename = "unspentTimestamp")]
+	pub unspent_timestamp: String,
+	#[serde(rename = "unspentRoots")]
+	pub unspent_roots: Vec<String>,
+	#[serde(rename = "unspentPathIndices")]
+	pub unspent_path_indices: String,
+	#[serde(rename = "unspentPathElements")]
+	pub unspent_path_elements: Vec<String>,
+	#[serde(rename = "spentTimestamp")]
+	pub spent_timestamp: String,
+	#[serde(rename = "spentRoots")]
+	pub spent_roots: Vec<String>,
+	#[serde(rename = "spentPathIndices")]
+	pub spent_path_indices: String,
+	#[serde(rename = "spentPathElements")]
+	pub spent_path_elements: Vec<String>,
+}
+
+#[derive(Debug)]
+pub struct RewardCircuitInputs {
+	pub rate: Vec<BigInt>,
+	pub fee: Vec<BigInt>,
+	pub reward_nullifier: Vec<BigInt>,
+	pub ext_data_hash: Vec<BigInt>,
+	pub note_chain_id: Vec<BigInt>,
+	pub note_amount: Vec<BigInt>,
+	pub note_asset_id: Vec<BigInt>,
+	pub note_token_id: Vec<BigInt>,
+	pub note_ak_x: Vec<BigInt>,
+	pub note_ak_y: Vec<BigInt>,
+	pub note_blinding: Vec<BigInt>,
+	pub note_path_elements: Vec<BigInt>,
+	pub note_path_indices: Vec<BigInt>,
+	pub note_alpha: Vec<BigInt>,
+	pub note_ak_alpha_x: Vec<BigInt>,
+	pub note_ak_alpha_y: Vec<BigInt>,
+	pub input_chain_id: Vec<BigInt>,
+	pub input_amount: Vec<BigInt>,
+	pub input_private_key: Vec<BigInt>,
+	pub input_blinding: Vec<BigInt>,
+	pub input_nullifier: Vec<BigInt>,
+	pub input_root: Vec<BigInt>,
+	pub input_path_elements: Vec<BigInt>,
+	pub input_path_indices: Vec<BigInt>,
+	pub output_chain_id: Vec<BigInt>,
+	pub output_amount: Vec<BigInt>,
+	pub output_private_key: Vec<BigInt>,
+	pub output_blinding: Vec<BigInt>,
+	pub output_commitment: Vec<BigInt>,
+	pub unspent_timestamp: Vec<BigInt>,
+	pub unspent_roots: Vec<BigInt>,
+	pub unspent_path_indices: Vec<BigInt>,
+	pub unspent_path_elements: Vec<BigInt>,
+	pub spent_timestamp: Vec<BigInt>,
+	pub spent_roots: Vec<BigInt>,
+	pub spent_path_indices: Vec<BigInt>,
+	pub spent_path_elements: Vec<BigInt>,
+}
+
+impl RewardCircuitInputs {
+	pub fn from_raw(inputs: &InputsRaw) -> Self {
+		let rate = vec![Self::to_bigint(&inputs.rate)];
+		let fee = vec![Self::to_bigint(&inputs.fee)];
+		let reward_nullifier = vec![Self::to_bigint(&inputs.reward_nullifier)];
+		let ext_data_hash = vec![Self::to_bigint(&inputs.ext_data_hash)];
+		let note_chain_id = vec![Self::to_bigint(&inputs.note_chain_id)];
+		let note_amount = vec![Self::to_bigint(&inputs.note_amount)];
+		let note_asset_id = vec![Self::to_bigint(&inputs.note_asset_id)];
+		let note_token_id = vec![Self::to_bigint(&inputs.note_token_id)];
+		let note_ak_x = vec![Self::to_bigint(&inputs.note_ak_x)];
+		let note_ak_y = vec![Self::to_bigint(&inputs.note_ak_y)];
+		let note_blinding = vec![Self::to_bigint(&inputs.note_blinding)];
+		let note_path_elements =
+			inputs.note_path_elements.iter().map(|val| Self::to_bigint(val)).collect();
+		let note_path_indices = vec![Self::to_bigint(&inputs.note_path_indices)];
+		let note_alpha = vec![Self::to_bigint(&inputs.note_alpha)];
+		let note_ak_alpha_x = vec![Self::to_bigint(&inputs.note_ak_alpha_x)];
+		let note_ak_alpha_y = vec![Self::to_bigint(&inputs.note_ak_alpha_y)];
+		let input_chain_id = vec![Self::to_bigint(&inputs.input_chain_id)];
+		let input_amount = vec![Self::to_bigint(&inputs.input_amount)];
+		let input_private_key = vec![Self::to_bigint(&inputs.input_private_key)];
+		let input_blinding = vec![Self::to_bigint(&inputs.input_blinding)];
+		let input_nullifier = vec![Self::to_bigint(&inputs.input_nullifier)];
+		let input_root = vec![Self::to_bigint(&inputs.input_root)];
+		let input_path_elements =
+			inputs.input_path_elements.iter().map(|val| Self::to_bigint(val)).collect();
+
+		let input_path_indices = vec![Self::to_bigint(&inputs.input_path_indices)];
+		let output_chain_id = vec![Self::to_bigint(&inputs.output_chain_id)];
+		let output_amount = vec![Self::to_bigint(&inputs.output_amount)];
+		let output_private_key = vec![Self::to_bigint(&inputs.output_private_key)];
+		let output_blinding = vec![Self::to_bigint(&inputs.output_blinding)];
+		let output_commitment = vec![Self::to_bigint(&inputs.output_commitment)];
+		let unspent_timestamp = vec![Self::to_bigint(&inputs.unspent_timestamp)];
+		let unspent_roots = inputs.unspent_roots.iter().map(|root| Self::to_bigint(root)).collect();
+		let unspent_path_indices = vec![Self::to_bigint(&inputs.unspent_path_indices)];
+		let unspent_path_elements =
+			inputs.unspent_path_elements.iter().map(|val| Self::to_bigint(val)).collect();
+		let spent_timestamp = vec![Self::to_bigint(&inputs.spent_timestamp)];
+		let spent_roots = inputs.spent_roots.iter().map(|val| Self::to_bigint(val)).collect();
+		let spent_path_indices = vec![Self::to_bigint(&inputs.spent_path_indices)];
+		let spent_path_elements =
+			inputs.spent_path_elements.iter().map(|val| Self::to_bigint(val)).collect();
+		Self {
+			rate,
+			fee,
+			reward_nullifier,
+			ext_data_hash,
+			note_chain_id,
+			note_amount,
+			note_asset_id,
+			note_token_id,
+			note_ak_x,
+			note_ak_y,
+			note_blinding,
+			note_path_elements,
+			note_path_indices,
+			note_alpha,
+			note_ak_alpha_x,
+			note_ak_alpha_y,
+			input_chain_id,
+			input_amount,
+			input_private_key,
+			input_blinding,
+			input_nullifier,
+			input_root,
+			input_path_elements,
+			input_path_indices,
+			output_chain_id,
+			output_amount,
+			output_private_key,
+			output_blinding,
+			output_commitment,
+			unspent_timestamp,
+			unspent_roots,
+			unspent_path_indices,
+			unspent_path_elements,
+			spent_timestamp,
+			spent_roots,
+			spent_path_indices,
+			spent_path_elements,
+		}
+	}
+	pub fn to_bigint(str_value: &str) -> BigInt {
+		match str_value {
+			hex_string if hex_string.starts_with("0x") =>
+				BigInt::from_bytes_be(Sign::Plus, &hex::decode(&hex_string[2..]).unwrap()),
+			decimal_string => BigInt::from_str(decimal_string).unwrap(),
+		}
+	}
+}
+
 
 pub fn deconstruct_public_inputs_reward_proof(
 	max_edges: usize,
